// src/models/types/enums.ts
export enum UserRankings {
    daily = 'daily',
    weekly = 'weekly',
    monthly = 'monthly',
    overall = 'overall',
  }
  
  export enum UserType {
    driver = 'driver',
    customer = 'customer',
    vendor = 'vendor',
  }
  
  export enum UserVendorType {
    foods = 'foods',
    items = 'items',
  }
  
  export enum UserVendorBusinessDays {
    sun = 'sun',
    mon = 'mon',
    tue = 'tue',
    wed = 'wed',
    thu = 'thu',
    fri = 'fri',
    sat = 'sat',
  }

  export enum MediaType {
    PROFILE_PHOTO = 'PROFILE_PHOTO',
    LICENSE = 'LICENSE',
    PASSPORT = 'PASSPORT',
    STUDENT_ID = 'STUDENT_ID',
    RESIDENT_PERMIT = 'RESIDENT_PERMIT',
    VENDOR_BANNER = 'VENDOR_BANNER'
}